moved most spark code to MRJobBinRunner
